pyproject.toml
[project]
name = "sf-annual-report"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "connectorx>=0.4.3",
    "numpy>=2.2.5",
    "pandas>=2.2.3",
    "polars<=1.28.0",
    "polars-ols>=0.3.5",
    "psycopg2-binary>=2.9.10",
    "pyarrow>=20.0.0",
    "python-dotenv>=1.1.0",
]
